Refactor to use new Endpoint base class
This changes the layer to use the new Endpoint base class for interface
layers, from juju-solutions/charms.reactive#123.  Some of this could be
cleaned up a bit more with changes to the charm, but I wanted to keep it
backwards compatible for now.

This also requires juju-solutions/charm-flannel#38, which is available
in revision 32 and higher of the flannel charm in the charm store, due
to the changes in the reactive lib.